Make the homepage more effecient
All pages with property cards could use the 800px version of the image instead of ever using the full version. Same with anecdotes, but the 480. These changes allow us to pass through a desired size to the responsive image template and get back only one size or the entire srcset.